GUI: Set the UI font
Accept notifications to set the GUI font from Neovim the notification name
name is "Gui", with the first argument "SetFont" followed by 0 or 1 additional
argument (the font name/size)

If not font name is passed it displays the current font, otherwise it sets the
current font e.g. "Monaco:h12" for 12pt Monaco. This follows the Vim font names
in Windows/Mac).

A command to set the font can be defined in nvimrc:

    :command -nargs=? Guifont call rpcnotify(0, 'Gui', 'SetFont', "<args>")

i.e. :Guifont Monaco:h12 can then be used to set the GUI font

When displaying the current font neovim-qt may print a different font than the
user set - this happens due to system font substitution, because it always
displays the actual font being used.

Fonts that do not exist, or are not fixed width cause an error message to
be displayed.