Resize Mainwindow when Neovim shell is resized
There are some upstream (Qt) issues for QWidget resizing (QTBUG-45086). That
make it harder to deal with Neovim resizes. For now the quickest way around
the issue seems to be wrapping the Shell widget in another widget (which we
already did) and propagate resizes to the parent widget.

Using the neovimResized() signal the mainwindow can adjust itself after a
shell resize event. For now the following behaviour is in effect:

1. If the window is maximised AND the shell is larger than the mainwindow,
   resize the shell to fit the window
2. If the window is maxmised and the shell is smaller than the window then,
   do nothing - this allows you to call :set lines while keeping the window
   maximized
3. If the window IS NOT maximised, then resize the window to the shell size.

Some minor refactoring to the Shell class:

- Split resize request code into a separate method Shell::resizeNeovim()
- Return int for neovim geometry sizes
- Added new signal Shell::neovimResized(), triggered when finishing a resize.