GUI: Don't call :qa on close for UNIX/TCP sockets
When the Neovim shell is closed (e.g. user closes window) the GUI would ignore
the event and call :qa, triggering the Neovim process to exit. This makes sense
when Neovim was launched by the GUI, since we need to cleanup the process.
However when connecting to an external Neovim process (TCP/UNIX socket) this
might not be what the user intends.

For TCP/UNIX sockets the "qa" command is no longer sent. This means that if the
user closes the GUI window the GUI will terminate the connection and close, but
the Neovim process will remain runnning. In cases where the user wants to exit
the Neovim process he would need to call :q himself.

This breaks the previous behaviour, but there was no way to override it. From a
user perspective the situation might improve later for UNIX sockets if the GUI
can discover available sockets.

Issue #36