GUI: Reject fonts with wrong italic/bold metrics
If we load a font that lacks bold/italic variants we risk it being
replaced by a font with incorrect metrics and causing drawing errors
(e.g. clipped characters). Instead refuse to load fonts that don't
have bold and italic variants with the same metrics.

This ends up being a bit too pedantic for users that don't need the
bold/italic fonts. The alternatives are:

- Use the regular font for bold/italic, but this would be confusing
  for the user - breaks expectations for schemes
- Synthesize bold/italic variants, but I have no plans to implement
  this directly since some systems already do it (e.g. fontconfig)