GUI: Support undercurl attribute
When drawing text the undercurl attribute was being represented as an
underline, and the underline attribute was ignored - this commit
corrects the use of underline and adds initial support for undercurl.

The undercurl is drawn over the text as a red dotted line, at the
bottom of the text cell. Currently the Neovim API does not expose
the guisp highlight attribute, so for now the color is hardcoded.