Raise fatal error on discoverMetadata() timeout
Non fatal communication errors may cause Neovim to fail to respond to
critical RPC requests. In such cases we need to enforce a timeout after
which the connecting client will fail. This commits enforces a timeout
of 10s for discoverMetadata(), if no response is received the
NeovimConnector raises a fatal error.