GUI: Set $WINDOWID in Neovim process
From issue #64 - in Neovim the v:windowid var is not available, while in
Unix a workaround is to use the $WINDOWID variable, in Windows there is
no alternative. This commit sets the $WINDOWID environment var inside
the attached Neovim process using a call to vim_command(let $WINDOWID="...").
The variable is only set when Neovim is spawned as a process.