Fix regression: initial screen was blank until user input was made
Commit 9ba51680e28376b3f58aae7489c8f8a3ad4d8400 introduced a bug that after the
application start you would only see a white screen with a broken cursor. It
would only disappear after some keyboard input.

I assume the reason for that was that we received notifications for drawing
events while waiting for the font information. The commit introducing this bug
says we need to wait with attaching until we receive the font, it didn't say
we can't connect the other signal handlers already, so this is what I did here.