fix default font on mac #86
Define a DEFAULT_FONT macro in shell.h that is set to "Courier New" on
Mac OS X and to "Monospace" on all other OSes. Use this macro to set the
default font in shell.cpp .

The reason for this is that on OS X "Monospace" creates some sort of
really horrible, ugly font.