fix ctrl and cmd on mac #86
Qt defines Meta as Control and Control as Cmd keys on OS X.

If that was the only problem, it would have been much easier to fix,
unfortunately for ctrl+key and cmd+key combos the text sent to the key
event is also empty, but the key code has the ascii/unicode value of the
key.

Define our own protected constants for the modifier keys, using
OS-specific #ifdefs so that they are mapped correctly on OS X as well as
some of the Key constants for the modifiers for use in input.cpp .

In convertKey in input.cpp, when text is empty, check if it's a ctrl+key
or cmd+key combo and not a ctrl or cmd key by itself, or not a
ctrl+alt+cmd combo of some sort by itself, and if so, use the key code
as the text.

In modPrefix in input.cpp, map the CmdModifier (which is ControlModifier
in Qt, mapped in input.h) to D- like MacVim does.

FIXME: cmd+key combos are mapped to <D-<key>> like MacVim does, but
there is no response to these key combos right now, perhaps this needs
to be fixed in neovim itself.

(NOTE: logic rewritten somewhat by maintainer)