GUI: Rewrite shell widget
Completely rewrite shell widget to store the full contents contents.
This helps with a number of outstanding bugs we have, and there is no
other way to fix them.

Likely this will cause some regressions, but I've kept the widget/store
implementation separate from the Neovim specific bits, hopefully this
makes it easier to move things forward.  The new widget implementation
is stored under src/gui/shellwidget. It has its own toplevel make file
and you can run it separately with its own unit tests, etc.

TODO: I have not worked on a proper resize implementation yet, right now
Neovim is only resized if the window is resized, but not the other way
around - we need to spend some time to figure out what is the expected
behaviour for resizes.