GUI: Fix cursor painting for wide characters
For #3, when drawing wide characters the chars takes up two cells
instead of one.

When drawing cells into the screen the cell that follows the position of
a wide char is not painted, because it was already painted in the
previous cell.

The cursor drawing also needs to take into consideration the size of the
cursor is wider for wide characters.