GUI: Improve widget resizing
Refactor: Handle the call to QWidget::updateGeometry() in the ShellWidget and call
QWidget::resizeEvent((). Instead of discarding concurrent calls
resizeNeovim, keep the last call and defer it until the previous call is
finished.

If the Shell widget is a top level window
1. If it is maximized or fullscreen, resize Neovim back, we cannot
   resize a maximized widget
2. Otherwise resize the widget.

If the shell widget is used as part of the MainWindow resize the
mainwindow when the Neovim shell widget is resized. For now the
maximized and fullscreen cases remain unimplemented.