GUI: Refactor MainWindow errors
Instead of using a DockWidget to display error messages use a
QStackedWidget to switch between the Shell and the ErrorWidget.

This should fix #84. The one downside of this change is that we
don't have anything resembling a "connecting" state in the main
window, sometimes when trying to connect to a closed socket a
flicker between the error and Shell widgets is clearly visible.