GUI: Handle Maximized/Fullscreen notifications
Process Gui WindowMaximized/WindowFullScreen notifications, to change
the GUI window state. Both notifications require an argument to indicate
the desired state (0 means disabled).

When the state changes the GUI sets g:GuiWindowMaximized and
g:GuiWindowFullScreen to 1 if enabled or 0 if disabled.

Depending on the Window Manager, these events might not be respected
meaning these variables will not match the expected window behaviour.