Add unit test for MainWindow Gui variables
There was already a unit test in tst_shell to check the assignment
of the Gui variables on startup, however this did not cover our
MainWindow widget. We now test both cases with the Shell widget and
the MainWindow.

For MainWindow we need to call updateGuiWindowState() after attaching to
Neovim.

There was some refactoring involved

- MainWindow now forwards the neovimAttached signal seen in the Shell
  class.
- To facilitate writing unit tests for the GUI a new library target
  was added - neovim-qt-gui - that holds all the reusable GUI
  components.
- nvim-qt and tst_shell now link against this library.