GUI: Normalize value of g:GuiFont
When the user sets the font, the GUI updates g:GuiFont to hold its
new value. However we were setting the value as passed by the user
which would allow for font descriptions that while valid can be a
bit ambiguous e.g. 'DejaVu Sans Mono:b:i:l', or are not accurate.
Normalize this value before setting it.