GUI: Accept filenames as positional arguments
Previously when spawning Neovim only arguments after "--" were forwarded
to Neovim. nvim-qt now takes filenames as positional parameters that are
appended to the Neovim arguments.

The following are equivalent and open three files

    nvim-qt -- file0 file1 file2
    nvim-qt file0 file1 file2

The following is also equivalent, but notice the order,
because positional arguments are appended at the end of
neovim arguments

    nvim-qt file1 file2 -- file0