GUI: New option --spawn
The --spawn option enables calling nvim using a customized
command, e.g.

    nvim-qt --spawn arg0 arg1 arg2 ...

This overrides all command flags that are usually passed
to nvim (e.g. --embed, set rtp, termguicolors). It is up
to the caller to make sure the arguments are correct.

This is meant as an alternative for cases when --nvim is
not enought, e.g. to execute nvim remotely using SSH

    nvim-qt --spawn ssh hostname nvim --embed

No other options except the command arguments may follow
--spawn. Qt 5.7 has an option in QCommandLineParser to
enable this behaviour, but for compatibility we implement
our own workaround here.