Avoid repeated setup for signal ::reconnectNeovim
The reconnect signal emited by the error dialog was being connect
with each call to MainWindow::init(), this meant that with each
reconnection the signal connected again, and emmited multiple times
for the same button push. One consequence was that the nvim process
was spawned multiple times as well.

Moved signal initialization to the MainWindow constructor.

- fixes #227