Add explicit check for keypad keys
Neovim expects keypad keys to be represented with the sequence <k0-9>
or <kHome> <kEnd>, etc.

In OSX the arrow keys are also considered as part of the key pad which
caused them to be encoded as <kUp> which is not a valid key for nvim.

Add an explicit check for the supported keys. I've allowed these keys
to have modifiers e.g. <S-k8> but in practice this seems to be system
dependent - for example in some systems Shift-keypad5 will generate no
key event, or Shift-keypad8 will be the Up arrow. Meaning that these
will never be accurately detected as keypad keys.