Use Q_ENUM, when available, for MsgpackError and NeovimError
Q_ENUM gives better feedback when checking the value of an enum in
tests.

FAIL!  : NeovimQt::Test::connectToNeovimTCP() Compared values are not the same
   Loc: [/«PKGBUILDDIR»/test/tst_neovimconnector.cpp(65)]

vs.

FAIL!  : NeovimQt::Test::connectToNeovimTCP() Compared values are not the same
   Actual   (c->errorCause())             : NoError
   Expected (NeovimConnector::SocketError): SocketError
   Loc: [/home/jamessan/neovim-qt-0.2.7/test/tst_neovimconnector.cpp(65)]

Q_ENUMS is [deprecated] in favor of Q_ENUM, introduced in Qt 5.5, but to
retain compatibility with older versions only use it when available.

[deprecated]: https://doc.qt.io/archives/qt-5.5/qobject-obsolete.html#Q_ENUMS