Add support nvim version runtime detection
The is a pretty big commit, the highlights

- dropped support for Neovim 0.1.5 and earlier (i.e. before API 1)**
- generate different bindings for different nvim versions

The rationale is that we need to do runtime detection of the nvim
version to enable some features.

Basically each version gets a class NeovimApiXX. And the NeovimConnector
provides getters for different versions. If one specific version is not
available it returns NULL.

To enable this the Function class no longer holds any references to
generated code, its just a function signature representation. Functions
are represented in calls using a 64bit integer, instead of enum we had
earlier.

We no longer match the runtime API metadata against the bindings, we
just look at the reported API version. The code for extensive
verifications is still there under NeovimApiXX::checkFunctions(). But
it is disabled in the connection code.

TODO:

- Add check for prerelease versions of nvim
- Support API version 0
- Add some more guards in the GUI shell code
- Move api version error out of the connector class and into the GUI
  code