Fix bug in EXT type unpacker: value was always 0
This will no doubt go into the hall of infamy of my mosty silly bugs.
The EXT unpacker function that decodes types Window,Tabpage,Buffer
deallocated the msgpack result BEFORE actually converting it to a
QVariant.

To make matters even worse the access to the data enum was made under
the assumption that an integer value was present, but since the data
was already cleared this value was always 0. The consequence of this was
that EXT types were ALWAYS decoded as 0.

- Updated all bindings with a fix, and added a type check
  that allows both signed and unsigned int.
- Added some tests for the various api versions using
  get_current_tabpage().