Fix wrong api version used in GUI
The ext_tabline feature requires api2, but I was using api1.

The deeper issue is that we have no mechanism to automatically fallback
to an earlier api version if attachment fails. And when it happens the
UI just blocks. So maybe we need a parameter to specify the api version
we want to use.

This commit simply moves up the requirement for api2. If api2 is not
available we use api0.