Implement clipboard provider in GUI
This commit does not alter default behaviour of the GUI, it only adds
the necessary RPC methods to get/set the clipboard.

- Vim (and Neovim) stores the selection type along with data in the
  clipboard. To achieve this the selection type (a string) is stored in
  the clipboard with the mimetype application/x-nvim-selection-type.
- In Windows/Mac both the * and + registers are the system clipboard

A clipboard provider is included in the test directory that forwards
clipboard operations to the gui. But we will have to push upstream to
support regular functions for the clipboard provider.

Also took the chance to

- remove the checkCommand in tst_shell.cpp, there was a race condition
  the tests that used it for the font commands
- Added some debugging and error reporting