Appveyor: Reenable tests in windows builds
For a while now the appveyor builds have been failling, or if they work
the tests will not run. This was caused by environment changes in the
appveyor images AND by an unknown runtime problem when executing the
tests.

As such I'm rewriting the appveyor yml file. For now I'm keeping the
msys2 static qt version out of the this, since its the most complex.