Windows: Enable HiDpi scalling
For now I'm setting this in windows only, there open
issues in upstream Qt that suggests this may have unexpected effects
in other systems (QTBUG-65061, QTBUG-65102), also QTBUG-63580 offers
contradictory information with what we have experienced.

The other way to do this is to set Qt::AA_EnableHighDpiScaling
but it does not seem to work on Windows. But it seems this does not
work on windows (#391).