GUI: Abort ongoing mode if input is blocked
Sometimes Neovim is blocked waiting for input. When this happens
only async functions can be called without blocking, at this time
these are

- nvim_input()
- nvim_get_mode()
- nvim_get_api_info()

One specific case of this is #425. Another is the late attachment of
the GUI when neovim is blocked on an error or waiting for the pager.

This commit adds a helper function that sends input Ctrl-C if neovim is
blocked. This is used to fix #425.