GUI: Include windowid in channel attributes
v:windowid (and g:GuiWindowId) hold the window id for the GUI. This
approach (v:windowid) comes from Vim, but in Neovim multiple UIs can be
attached to the same nvim instances.

To address this we can include the window id in the channel information.
The addition of an api to grab this information is left for a later
time.