Cmdwidget : fix appearance of cmd_lines widgets
The command lines were losing space because of a frame and a vertical
scrollbar. Now both of these are off.

I introduced a magic number (which will need to be taken care of) to get
the characters to fit (in height) inside their command_line. This magic
number needed tweaks to the anchor_y value too in order to get the
effect we wanted (command line is flushed with the bottom of the `Shell`
widget)