Cmdline : Fix cursor
The cursor only appear when the widget has the focus. So `CmdLine`
takes focus whenever it computes a line. This way we're ensured the
cursor will always be in the line corresponding to the most recent
`cmdline_show` level.

Also, having the cursor appear allowed to fix bad offset when setting
cursor position from a `cmdline_show` or a `cmdline_pos` call.