Fix wrong row count in shellwidget
In the shellwidget the paint event rounds down the row/column to comply
to the actual shell dimensions. Turns out the statment that handles the
rows is actually readjusting the columns - this has two potential
consequences

1. we could be painting wrong data below the last row
2. we could be painting areas (columns) that do not need painting

I was unable to reproduce cases of 1. It should be noted that because
the widget always resizes the shell when a resize occurs, the difference
between the last row/column would always be fractional.