Adding a context menu with Cut/Copy/Paste operations
This change modifies MainWindow to add Cut/Copy/Paste/SelectAll events. All of
these events are appended to a context menu which is triggered on right-click.

The function Shell::neovimMouseEvent(...) is modified to trigger right click
events on right mouse button release. Some of these event may no longer be
passed along to the NeoVim backend.