GuiPopupMenu size logic modifications and code cleanup
The behavior between NeoVim TUI PUM (popupmenu) and NeoVim-Qt PUM are
different. Based on user feedback from #524 the current logic may be confusing
to users. This change modifies the NeoVim-Qt PUM show logic to be as similar
as possible to TUI, while remaining simple.

In addition to the logic changes, the code has been cleaned and modularized
while making these changes.