shell: fix duplicate cursors in terminal mode
In terminal mode, nvim draws the cursor itself by writing space
characters with the "reverse" attribute.  The only apparent signal
to the UI is that "busy" is turned on until terminal mode is left.
The UI API documentation recommends not drawing the cursor while
nvim is busy, which is how double cursors are avoided by other
UIs.  Do the same thing.