Improve window size restore cmdline argument handling
The previous window size is not restored when '--fullscreen', '--geometry', or
'--qwindowgeometry' arguments are present. This prevents size restoration code
from overriding user-provided settings.

Saving window state is disabled for sessions started with '--fullscreen' this
prevents the annoying case of subsequents sessions defaulting to fullscreen
mode.

Note: '--maximized' is explicitly NOT excluded. We may still want to restore
the window size when this flag is provided, as the window will "remember" its
last size when un-maximized.