QCommandLineParser geometry argument improvements
Issue #587: QCommandLineParser issues a qWarning when isSet is called for an
argument that has not been defined. This is fixed by adding platform specific
logic to only check the argument which exists --geometry or --qwindowgeometry.

Issue #565 The help message on some platforms displays the wrong --geometry
argument on Windows/Apple platforms. The fix for Issue 587 is used to add the
correct help message.