Add support for Neovim guicursor options
Extends ShellWidget to support a basic cursor abstraction. When the Shell
object receives style updates, it can forward this styling information.
ShellWidget will then handle the rendering of the cursor (color, blink state,
position, etc).

The guicursor option is on-by-default, and can be disabled via:
  `:set guicursor=`

This will cause neovim-qt to fall back to the old XOR based cursor.