Issue 628: GuiFont! fails for bad monospace fonts
The line following line from setAttaced is invalid:
	`m_nvim->api0()->vim_set_option("guifont", fontDesc());`

This works for well formed monospace fonts, but does not work for bad
monospace fonts. This line updates the value of `guifont` with the default
font (valid monospace) on initialization.

If the user then sets a bad monospace font with GuiFont!, in some cases timing
of the neovim guifont events can override the user defined font.