Issue 648: IncSearch not rendered properly with ext_linegrid.
With `ext_linegrid` some highlights do not render properly:
QMap(("reverse", QVariant(bool, true)))

Any highlight that uses the default colors and "reverse" does not render in
reverse.

This occurs because the reverse parameter is not actually passed to the
rendering layer in ShellWidget.

Adds the "reverse" flag, so default colors can be determined at render time.
This is necessary to support `set background`.