WIP: Make Gui Close a synchronous request
If nvim sends a notification for closing it may exit before the
notification is handled. Fix this by supporting a synchronous request
for Gui Close.

The shim was updated to notify all available UIs.

TODO: this requires a version check - nvim_list_uis() might not be
available. The call to rpcrequest may fail and needs to checked for
error and fall back to the notification.