nvim-qt --help/version should work without $DISPLAY
The intent is to split command line handling from the App instance, so
that we can do it using a QCoreApplication and QCommandLineParser. To
achieve this i made multiple functions in the App class static.

Added a small test for #707 i.e. the following CLI invocations should
work without a $DISPLAY variable:

    --version
    --help

I've left --help-all out of this, since I cant get it to behave
reliably.