Add Font Ligature Support
This work would not have been done without @Nucearo 's efforts:
        https://github.com/equalsraf/neovim-qt/pull/470

Uses QTextLayout to render segments of similarly styled cells. QTextLayout
provides greater control over Qt's font-rendering engine via QGlyphRun. With
QGlyphRun, we are able to identify individual characters and their positions.

QGlyphRun should provide a means of correcting the spacing for non-perfect
monospace fonts. It also allows for identification of when a character
sequence is converted to a glyph.