Set initial editor size properly
When you run `nvim-qt -- -O a b` in your console, you'll see inequally
divided two editor pains in the opened Neovim-Qt. This is because the
initial editor size (lines and columns) is not based on the actual
window size.

To resolve this issue, we need to know the window size and cell size
before initializing Neovim. However it's impossible with the current
implementation, because the window is shown only after Neovim
initialization (delayed show mechanism first introduced in #70). I tried
to see if I could get Qt to calculate the layout before the window is
shown, but I couldn't find a way.

Therefore this commit tweaks three things:

- Show window (and shell widget) first and attach to neovim after the
  main window is shown.
- Use actual window size to calculate initial attaching size instead of
  the hard-coded value (the current code is assuming 60% of display size
  as a initial height and width, regardless of the actual window size).
- Remove delayed show mechanism entirely.

In addition to those, prevent init() from being called multiple times.
This is because init() is now called from showEvent() but showEvent() is
emitted not only when the window is newly created but also minimized
window is restored.