Fix MacOS CI Break
Travis displays the following error on MacOS:
$ brew ls | grep -wq cmake || brew install cmake
Error: Calling `brew list` to only list formulae is disabled! Use `brew list --formula` instead.

According to:
https://discourse.brew.sh/t/brew-list-less-fails/9224
"brew list would only list formulae when the output is not to a terminal."

We should add the --formula argument as suggested, this will keep the same
behavior as before the brew terminal API change.