Issue 806: Remove hard coded NVIM_QT_RUNTIME_PATH
Hard coding the compile-time install path is sub-optimal.

This can cause issues when the intended install location is changed. For
example the program is compiled on Windows with D:\Program Files. The
users installs to C:\Program Files.

This can impact Linux too, but is less likely: extract package image to /opt.

This has already been removed for MacOS. Removing for Linux/Windows
allows us to use the same code for all platforms.