Issue 817: Change Ctrl + Shift modifier handling
The Ctrl + Shift handling logic is not what nvim expects.

Previously:
  CTRL + B: <C-b>
  CTRL + SHIFT + B: <C-B>

New:
  CTRL + B: <C-b>
  CTRL + SHIFT + B: <C-S-B>

The logic has been changed so that if <C- or <A- are present, we will preserve
this shift key. This should prevent SHIFT + B from sending incorrect logic
such as <S-B>, while preserving all modifiers on command sequences.