Refactor tests to avoid QPair brace initializer
Building against Qt6 with GCC 10 fails for existing initialization
of QList<QPair<QKeyEvent, QString>>. QPair in qt6 is a typedef for
std::pair.

To avoid the use of QPair, move all the test items into a single class.
Instead of storing a QKeyEvent store its components (type, key,
modifiers) and build the final QKeyEvent when needed.