Fix :256cq test cases
On UNIX-like operating systems, an exit code is represented by a 8 bit
unsigned integer. So 256 overflows and appears to be 0 to the caller.
For reference, we can see man for wait(2): WEXITSTATUS() returns the
least significant 8 bits of the status returned by the wait() system
call.

On Windows, however, the exit code is known to be a signed 32-bit
integer. Therefore 256 doesn't cause overflow and it appears to be 256
as it is.