Refactor connection creation
On macOS, OS provides a special switching (Cmd+`) for the windows in the same
process. In order to properly support multiple windows, we need to
decouple the connection from the app so that the connection can be owned
by the window and the proper clean up happens when the window exits.

This code doesn't change any behavior since there's still only one
window per app.