Add ClangFormatDiff GitHub Action
This is a new CI rule that checks Pull Requests for clang-format
compliance. It will only apply to changed lines in Pull Requests.

The rule operates by analyzing the git diff since the PR fork-point. The
computed diff is piped into clang-format-diff.py, and the CI fails if
any clang-format differences are detected.

The PR author can apply the results:
 - Downloading a clang_format.patch Artifact
 - Adding a "auto-commit" label, commits will be auto-applied
 - Locally running contrib/clang-format-diff.sh (Linux/Mac).

This rule will only run for PRs, so we won't get noise for violations of
the existing code. We can also safely ignore clang-format, if manual
formatting is desired.