Set showtabline in GuiTabline
Context:

Displaying tabs in neovim is controlled by the option showtabline:
* 0 -> never display
* 1 -> display if more than 1 tab (default)
* 2 -> always display

Ostensibly, #739 (Improved GuiTabline show buffers) changed the behavior of
neovim such that buffers will be displayed in the tabline when GuiTabline is
set, but showtabline's default makes GuiTabline a no-op until a second neovim
tab is opened (with :tabnew, etc). The fix is to save the showtabline value and
then when calling GuiTabline, if enabled set showtabline to 2, otherwise reset
to the saved value.

This enables the (somewhat silly) configuration of setting showtabline to 0 and
then toggling GuiTabline as the only means of seeing the tabline.