GuiTabline and showtabline behavior
Previously, GuiTabline used tab count only for the showtabline option.
For the default value of showtabline=1 (At Least Two), this meant the
GuiTabline was never displayed for multiple buffers. This is not
intuitive.

Change GuiTabline to evaluate buffers or tabs for the showtabline
count: two buffers and no tabs will now cause GuiTabline to show.

See Pull Request 973 for more details