Fix Unit Tests
The previous commit changes the way Shell is displayed, and breaks some
of the existing tests because signal neovimAttachmentChanged does not
fire the same as before: now `->show()` is required.

Disables the local ginit.vim file, to prevent local configuration from
impacting the tests. This also prevents error messages that do not fit
on the screen from triggering a 'Press Enter' prompt and test failure.