Fix GuiFont test
We are observing a failing test for font changes

1. the finished signal for the call to vim_command is not caught by the
   signal spy

The root of the issues seems to be a failure to set the font due to some
combination of bad font metrics happening in the test environment. The
GuiFont command was changed to force the change. The timeout was also
increased.