Issue 1063: change default guicursor behavior
Currently, when no guicursor value is provided (styling is disabled), we
use some default values: block, ver25, or hor25 based on mode.

The TUI does not behave the same way, it uses block mode for all cases
when guicursor is empty.

This change removes the cursor defaults, and applies block mode to all
cases where guicursor is empty.