settings: store window geometry in ~/.config/nvim-qt/window_geometry.conf
The single-argument QSettings constructor sets the organization name.
nvim-qt is passing "window-geometry" as the organization name, which
results in QSettings using `~/.config/window-geometry.conf`.
This is undesirable because the file is written to the root of
`~/.config` without any indication about which application created
or uses the file.

Set the organization name to "nvim-qt" when writing window geometry
so that `~/.config/nvim-qt/window-geometry.conf` is used instead.