Set a default size on first launch
When `~/.config/nvim-qt/window-geometry.conf` does not exist then
restoreGeometry() will initialize a window that is tiny and must be
resized in order to be usable in some environments.

Resize the window to half the screen when launched for the first time.
Do not restore window state when it is invalid or missing.