Try to make builds reproducible
Reproducible builds for neovim-qt are not working because mapping flags
passed to the cmake build are included in the code (through version.h.in).
Meaning that builds with slightly different mapping flags end up being
different; even if no other compiler flags are different.

ref https://github.com/equalsraf/neovim-qt/issues/1111

Adds build option for reproducible builds. Currently only affects the use
of compilation flags in the --version output.