Update msgpack(-c) to 6.1.0 and simplify fetching it
msgpack split its mixed C/C++ tarball into two separate ones starting
with 4.0 [1][2]. Since 6.0.0 the C part uses msgpack-c as a name for
everything. This is where the requirement for 6.0.0 comes from, but it
has been released over two years ago and I'd expect it to be available
for everybody who wants to use -DUSE_SYSTEM_MSGPACK=TRUE.
It allows to just use the imported target from msgpack-c's cmake config
file and to simplify the cmake code a bit.
This fixes the build with CMake >= 4.0, which has removed compatibility
with CMake < 3.5 while msgpack 3.2.0 sets
"CMAKE_MINIMUM_REQUIRED (VERSION 2.8.12)".
Also raise the cmake requirement to 3.11 for FetchContent.

[1] https://github.com/msgpack/msgpack-c/pull/876
[2] https://github.com/msgpack/msgpack-c/pull/878
[3] https://github.com/msgpack/msgpack-c/pull/1053